requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.3
requests==2.32.3
langchain==0.3.0
langchain-openai==0.2.0

